language model: fix model serialization
Distiller is not explicitly used by the generate.py script, but because main.py
serializes the entire model (look for 'torch.save(model, f)' in main.py), it also creates a
dependency on distiller code.
It's a bit ironic that PyTorch's docs advise against this kind of serialization,
while PyTorch's samples use it: https://pytorch.org/docs/master/notes/serialization.html

For now, because we don't want to make too many changes in the example code that comes from
PyTorch, we will work around this by importing Distiller.  In the future, we may change
the serialization code.